Backend/document_analyzer/pyproject.toml
[project]
name = "document-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "Flask",
  "flask-cors",
  "moviepy",
  "python-docx",
  "pypdf",
  "python-pptx",
  "google-generativeai",
  "requests",           # Likely needed for API calls
  "pydub",              # For audio processing
  "SpeechRecognition",  # If you're using mic/audio input
  "whisper",            # If using OpenAI Whisper
  "torch",              # Required for Whisper
  "openai"              # May be used in gemini_call fallback
]
